Address review: free-text namespace, move target impl into v2, client eviction
- Namespace picker is now an editable combobox with free-text entry, so users
  on RBAC-restricted clusters that deny list-namespaces can type their
  namespace by hand; an inline error is shown when listing fails (defaults
  still fall back to the current namespace).
- Move the cluster target enumeration out of the wizard module into
  ui/server/v2/targets.rs so v2 owns it instead of calling across modules;
  the wizard module keeps only its is-returning state.
- Reuse the existing UiResult alias in v2 instead of a duplicate V2Result.
- Evict a cached kube client when a request made with it fails, so a stale
  client (rotated CA, moved API server, broken auth-exec) is rebuilt rather
  than reused.
- Fix broken rustdoc intra-doc links and full-workspace rustfmt.
